Fix the order of boxes layers
the "back to home" button doesn't work on the desktop screen. (means that isn't clickable).
the blue shape on the left is on top of the box where the button is.
Add z-index: -1; to the .design-left class which represents the blue shape background in the left. (You can also change the order in the HTML (Advised) )